reorganized files of TimeInput
